Do not fetch the description when on list view
